requirements.txt
Django==3.2.4
google-api-python-client==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4